+++ h2/src/tools/org/h2/dev/util/FileViewer.java
@@ -57,13 +57,13 @@ public class FileViewer extends Tool {
             } else if (arg.equals("-find")) {
                 find = args[++i];
             } else if (arg.equals("-start")) {
-                start = Long.decode(args[++i]).longValue();
+                start = Long.decode(args[++i]);
             } else if (arg.equals("-head")) {
                 head = true;
             } else if (arg.equals("-tail")) {
                 tail = true;
             } else if (arg.equals("-lines")) {
-                lines = Integer.decode(args[++i]).intValue();
+                lines = Integer.decode(args[++i]);
             } else if (arg.equals("-quiet")) {
                 quiet = true;
             } else if (arg.equals("-help") || arg.equals("-?")) {
